@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -29,6 +29,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   http: ^0.13.3
   get: ^4.1.4
+  dio: ^4.0.0
 
 dev_dependencies:
   flutter_test:
