@@ -1,23 +1,23 @@
 class AppException implements Exception {
-  final String message;
-  final String prefix;
-  final String url;
+  final String? message;
+  final String? prefix;
+  final String? url;
 
   AppException([this.message, this.prefix, this.url]);
 }
 
 class BadRequestException extends AppException {
-  BadRequestException([String message, String url]) : super(message, 'Bad Request', url);
+  BadRequestException([String? message, String? url]) : super(message, 'Bad Request', url);
 }
 
 class FetchDataException extends AppException {
-  FetchDataException([String message, String url]) : super(message, 'Unable to process', url);
+  FetchDataException([String? message, String? url]) : super(message, 'Unable to process', url);
 }
 
 class ApiNotRespondingException extends AppException {
-  ApiNotRespondingException([String message, String url]) : super(message, 'Api not responded in time', url);
+  ApiNotRespondingException([String? message, String? url]) : super(message, 'Api not responded in time', url);
 }
 
 class UnAuthorizedException extends AppException {
-  UnAuthorizedException([String message, String url]) : super(message, 'UnAuthorized request', url);
+  UnAuthorizedException([String? message, String? url]) : super(message, 'UnAuthorized request', url);
 }
