@@ -3,8 +3,7 @@ import 'package:get/get.dart';
 
 class DialogHelper {
   //show error dialog
-  static void showErroDialog(
-      {String title = 'Error', String description = 'Something went wrong'}) {
+  static void showErroDialog({String title = 'Error', String? description = 'Something went wrong'}) {
     Get.dialog(
       Dialog(
         child: Padding(
@@ -13,7 +12,7 @@ class DialogHelper {
             mainAxisSize: MainAxisSize.min,
             children: [
               Text(
-                title ?? '',
+                title,
                 style: Get.textTheme.headline4,
               ),
               Text(
@@ -22,7 +21,7 @@ class DialogHelper {
               ),
               ElevatedButton(
                 onPressed: () {
-                  if (Get.isDialogOpen) Get.back();
+                  if (Get.isDialogOpen!) Get.back();
                 },
                 child: Text('Okay'),
               ),
@@ -36,7 +35,7 @@ class DialogHelper {
   //show toast
   //show snack bar
   //show loading
-  static void showLoading([String message]) {
+  static void showLoading([String? message]) {
     Get.dialog(
       Dialog(
         child: Padding(
@@ -56,6 +55,6 @@ class DialogHelper {
 
   //hide loading
   static void hideLoading() {
-    if (Get.isDialogOpen) Get.back();
+    if (Get.isDialogOpen!) Get.back();
   }
 }
