@@ -2,16 +2,17 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:http/http.dart' as http;
+import 'package:dio/dio.dart';
+
 import 'app_exceptions.dart';
 
-class BaseClient {
+class DioClient {
   static const int TIME_OUT_DURATION = 20;
   //GET
   Future<dynamic> get(String baseUrl, String api) async {
     var uri = Uri.parse(baseUrl + api);
     try {
-      var response = await http.get(uri).timeout(Duration(seconds: TIME_OUT_DURATION));
+      var response = await Dio().get(baseUrl + api).timeout(Duration(seconds: TIME_OUT_DURATION));
       return _processResponse(response);
     } on SocketException {
       throw FetchDataException('No Internet connection', uri.toString());
@@ -25,9 +26,7 @@ class BaseClient {
     var uri = Uri.parse(baseUrl + api);
     var payload = json.encode(payloadObj);
     try {
-      var response = await http.post(uri, body: payload).timeout(Duration(seconds: TIME_OUT_DURATION));
-      throw BadRequestException(
-          '{"reason":"your message is incorrect", "reason_code":"invalid_message"}', response.request!.url.toString());
+      var response = await Dio().post(baseUrl + api, data: payload).timeout(Duration(seconds: TIME_OUT_DURATION));
       return _processResponse(response);
     } on SocketException {
       throw FetchDataException('No Internet connection', uri.toString());
@@ -39,23 +38,16 @@ class BaseClient {
   //DELETE
   //OTHER
 
-  dynamic _processResponse(http.Response response) {
+  dynamic _processResponse(response) {
     switch (response.statusCode) {
       case 200:
-        var responseJson = utf8.decode(response.bodyBytes);
-        return responseJson;
-        break;
       case 201:
-        var responseJson = utf8.decode(response.bodyBytes);
-        return responseJson;
-        break;
+        return response.data;
       case 400:
         throw BadRequestException(utf8.decode(response.bodyBytes), response.request!.url.toString());
       case 401:
       case 403:
         throw UnAuthorizedException(utf8.decode(response.bodyBytes), response.request!.url.toString());
-      case 422:
-        throw BadRequestException(utf8.decode(response.bodyBytes), response.request!.url.toString());
       case 500:
       default:
         throw FetchDataException('Error occured with code : ${response.statusCode}', response.request!.url.toString());
